pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "envsync"
version = "0.1.0"
description = "CLI tool to detect configuration drift across environments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.24.0,<1.0.0",
  "PyYAML>=6.0.0,<7.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.0,<10.0.0"
]

[project.scripts]
envsync = "envsync.cli.app:app"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["envsync*"]
